Refaktorering av VedtakOgBeregning Begynt på AvslåVedtak
