+++ src/frontend/komponenter/Behandling/VedtakOgBeregning/InnvilgeVedtak/InntektsperiodeValg.tsx
@@ -1,16 +1,20 @@
 import { Element } from 'nav-frontend-typografi';
-import { EInntektsperiodeProperty, IBeløpsperiode, IInntektsperiode } from '../../../typer/vedtak';
+import {
+    EInntektsperiodeProperty,
+    IBeløpsperiode,
+    IInntektsperiode,
+} from '../../../../typer/vedtak';
 import { AddCircle, Delete } from '@navikt/ds-icons';
 import { AlertStripeFeil } from 'nav-frontend-alertstriper';
 import { Textarea } from 'nav-frontend-skjema';
-import MånedÅrVelger from '../../Felleskomponenter/MånedÅr/MånedÅrVelger';
+import MånedÅrVelger from '../../../Felleskomponenter/MånedÅr/MånedÅrVelger';
 import { Flatknapp } from 'nav-frontend-knapper';
 import React from 'react';
 import styled from 'styled-components';
-import InputMedTusenSkille from '../../Felleskomponenter/InputMedTusenskille';
-import { harTallverdi, tilTallverdi } from '../../../utils/utils';
+import InputMedTusenSkille from '../../../Felleskomponenter/InputMedTusenskille';
+import { harTallverdi, tilTallverdi } from '../../../../utils/utils';
 import Utregningstabell from './Utregningstabell';
-import { Ressurs } from '../../../typer/ressurs';
+import { Ressurs } from '../../../../typer/ressurs';
 
 const Knapper = styled.div`
     max-width: 500px;
