+++ src/frontend/komponenter/Behandling/VedtakOgBeregning/InnvilgeVedtak/Utregningstabell.tsx
@@ -1,9 +1,9 @@
-import { IBeløpsperiode } from '../../../typer/vedtak';
+import { IBeløpsperiode } from '../../../../typer/vedtak';
 import React from 'react';
-import DataViewer from '../../Felleskomponenter/DataViewer/DataViewer';
-import { Ressurs } from '../../../typer/ressurs';
-import TabellVisning, { TabellIkon } from '../TabellVisning';
-import { formaterNullableMånedÅr, formaterTallMedTusenSkille } from '../../../utils/formatter';
+import DataViewer from '../../../Felleskomponenter/DataViewer/DataViewer';
+import { Ressurs } from '../../../../typer/ressurs';
+import TabellVisning, { TabellIkon } from '../../TabellVisning';
+import { formaterNullableMånedÅr, formaterTallMedTusenSkille } from '../../../../utils/formatter';
 
 interface Props {
     beregnetStønad: Ressurs<IBeløpsperiode[]>;
